Use Powershell rules for globbing files.
Instead of trying to combine typical globbing rules with Powershell,
just rely on exclude parameter.

Remove SourceDirectory as a parameter, files can only be specified
with Files as an array of strings.

If a directory is specified, or a wildcard under a directory, upload
that directory exactly to destination, with name included. If a file
is specified, file goes directly to destination.